pyproject.toml
[project]
name = "scuba-doo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "supabase>=2.27.1",
    "uvicorn>=0.40.0",
]
